Rename verify to verifier and move it to internal
